@@ -13,7 +13,7 @@ async function main (){
     let startMs = Date.now()
 
 
-    const key = await dataparty_crypto.Routines.createKeyFromPasswordArgon2(
+    const key = await dataparty_crypto.Identity.fromPasswordArgon2(
         argon2,
         "supersecretpassword123",
         salt
@@ -24,7 +24,6 @@ async function main (){
 
     console.log('key')
     console.log(key)
-
 
     const deltaMs = endMs - startMs
 
