@@ -1,15 +1,46 @@
-
 import { Buffer } from 'buffer'
+
+import * as base64 from "@stablelib/base64";
+
 import Message from "./message";
-import { createKey, createKeyFromMnemonic, createKeyFromPassword } from "./routines";
+
+import {
+  reach,
+  getRandomBuffer,
+  createKey,
+  createSeedFromMnemonic,
+  createSeedFromPasswordPbkdf2,
+  createSeedFromPasswordArgon2,
+  createPQSharedSecret,
+  recoverPQSharedSecret,
+  createNaclSharedSecret,
+  createAESStream
+} from "./routines";
+
+
 
 export default class Identity implements IIdentity {
   id: string;
   key: IKey;
+  seed?: Buffer;
 
   constructor(opts = {} as any) {
     this.id = opts.id || "";
-    this.key = !opts || !opts.key ? createKey() : opts.key;
+    this.key = !opts || !opts.key ? null : opts.key;
+    this.seed = !opts || !opts.seed ? null : opts.seed;
+  }
+
+  async initialize(){
+    if(this.key != null && this.key.private){
+      throw new Error('identity already initialized')
+    }
+
+    if(!this.seed){ this.seed = await getRandomBuffer(64) }
+
+    let type = undefined
+    if(this.key && this.key.type){ type = this.key.type }
+
+    this.key = await createKey(this.seed, true, type)
   }
 
   async encrypt(msg, to :IIdentity) {
@@ -37,26 +68,120 @@ export default class Identity implements IIdentity {
     return message.verify(this);
   }
 
+  assertHasPostQuatumKEM(){
+    if(!this.hasPostQuatumKEM()){ throw new Error('no post quantum KEM') }
+  }
+
+  hasPostQuatumKEM(){
+    return reach(this.key, 'public.pqkem', null) !== null
+  }
+
+  async createStream(
+    to: IIdentity,
+    requirePostQuantum: boolean = true,
+    info?: Uint8Array | string,
+    salt?: Uint8Array | string
+  ) : Promise<IAESStreamOffer> {
+  
+    let pqSharedSecret = null
+
+    if(requirePostQuantum){
+      to.assertHasPostQuatumKEM()
+      this.assertHasPostQuatumKEM()
+    }
+    
+    if(this.hasPostQuatumKEM() && to.hasPostQuatumKEM()){
+      
+      pqSharedSecret = await createPQSharedSecret(to)
+      
+    }
+    
+    const naclSharedSecret = await createNaclSharedSecret(to, this)
+    const streamNonce = await getRandomBuffer(12)
+
+
+    console.log('aes-stream pqss', pqSharedSecret)
+    console.log('aes-stream naclss', naclSharedSecret)
+    
+    const stream = await createAESStream(
+        naclSharedSecret,
+        pqSharedSecret,
+        streamNonce,
+        info,
+        salt
+    )
+    
+    return {
+      sender: this.publicIdentity(),
+      pqCipherText: pqSharedSecret==null ? null : pqSharedSecret.cipherText,
+      streamNonce: base64.encode(streamNonce),
+      stream
+    }
+  }
+
+  async recoverStream(
+    offer: IAESStreamOffer,
+    requirePostQuantum: boolean = true,
+    info?: Uint8Array | string,
+    salt?: Uint8Array | string
+  ) : Promise<IAESStream> {
+
+    let pqSharedSecret = null
+
+    if(requirePostQuantum){
+      offer.sender.assertHasPostQuatumKEM()
+      this.assertHasPostQuatumKEM()
+    }
+    
+    if(this.hasPostQuatumKEM() && offer.sender.hasPostQuatumKEM()){
+      
+      pqSharedSecret = await recoverPQSharedSecret(this, offer.pqCipherText)
+      
+    }
+
+    const naclSharedSecret = await createNaclSharedSecret(offer.sender, this)
+
+    console.log('recover-stream pqss', pqSharedSecret)
+    console.log('recover-stream naclss', naclSharedSecret)
+
+    const stream = await createAESStream(
+      naclSharedSecret,
+      pqSharedSecret,
+      base64.decode(offer.streamNonce),
+      info,
+      salt
+    )
+
+    return stream
+
+  }
+
+  publicIdentity(){
+    return Identity.fromString( JSON.stringify(this.toJSON(false)) )
+  }
+
   /**
    *
    * @param extract if true, remove private key
    */
-  toJSON(extract?: boolean) {
+  toJSON(extract: boolean = false) {
     return {
       id: this.id,
+      seed: extract==true ? this.seed :  undefined,
       key: {
         type: this.key.type,
+        hash: this.key.hash,
         public: this.key.public,
-        private: !extract ? undefined : this.key.private
+        private: extract == true ? this.key.private : undefined
       }
     };
   }
 
   toMini() {
     return {
-      id: this.id,
-      public: this.key.public,
-      type: this.key.type
+      hash: this.key.hash,
+      type: this.key.type,
+      public: this.key.public
     };
   }
 
@@ -69,19 +194,60 @@ export default class Identity implements IIdentity {
     return new Identity(parsed);
   }
 
-  static fromMnemonic(phrase: string) {
-    return new Identity({
-      key: createKeyFromMnemonic(phrase)
-    });
+  static async fromRandomSeed(opts: any={}){
+    const seed = await getRandomBuffer(64)
+    const key = await createKey(seed)
+
+    return new Identity({ key, seed, ...opts })
   }
 
-  static fromPassword(
+  static async fromMnemonic(phrase: string, password: string, argon2: any, opts: any={}) {
+
+    const seed = await createSeedFromMnemonic(phrase, password, argon2)
+
+    console.log('seed type', typeof seed)
+
+    const key = await createKey( seed )
+
+    return new Identity({ key, seed, ...opts })
+  }
+
+  static async fromPasswordPbkdf2(
     password: string,
     salt: Buffer,
-    rounds?: Number
+    rounds?: number,
+    opts: any={}
   ) {
-    return new Identity({
-      key: createKeyFromPassword(password, salt, rounds)
-    });
+    const seed = await createSeedFromPasswordPbkdf2(password, salt, rounds)
+    const key = await createKey( seed )
+
+    return new Identity({ key, seed, ...opts })
+  }
+
+  static async fromPasswordArgon2(
+    argon: any,
+    password: string,
+    salt?: Uint8Array,
+    timeCost?: Number,
+    memoryCost?: Number,
+    parallelism?: Number,
+    type?: string,
+    hashLength?: Number,
+    opts: any={}
+  ) {
+    const seed = await createSeedFromPasswordArgon2(
+      argon,
+      password,
+      salt,
+      timeCost,
+      memoryCost,
+      parallelism,
+      type,
+      hashLength
+    )
+
+    const key = await createKey( seed )
+
+    return new Identity({ key, seed, ...opts })
   }
 }
