@@ -2,7 +2,8 @@
 const dataparty_crypto = {
   Routines: require('./routines'),
   Message: require('./message').default,
-  Identity: require('./identity').default
+  Identity: require('./identity').default,
+  PQIdentity: require('./pq-identity').default
 }
 
 
